backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Vector Store
chromadb==0.4.24
numpy<2.0

# LLM and Embeddings
openai==1.12.0
google-generativeai==0.3.2

# Document Processing
pymupdf==1.23.8
python-docx==1.1.0

# Web Search
requests==2.31.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
uuid6==2024.1.12

# CORS
starlette==0.35.1
